requirements.txt
appdirs==1.4.0
clang==3.8
common==0.1.2
dj-static==0.0.6
Django==1.9.6
django-admin-tools==0.7.2
django-appconf==0.6
django-bootstrap==0.2.4
django-bower==5.1.0
django-common==0.1.51
django-compressor==1.6
django-crispy-forms==1.6.0
django-extensions==1.6.7
django-select-multiple-field==0.3.0
django-sortedm2m==1.3.0
djangorestframework==3.3.3
docopt==0.6.2
drf-nested-routers==0.9.0
grunt==0.1.1
gunicorn==19.1.1
image==1.5.3
Naked==0.1.31
node==0.9.16
npm==0.1.1
numpy==1.11.0
odict==1.5.1
optional-django==0.1.0
pbr==1.10.0
Pillow==3.2.0
plumber==1.3.1
pycparser==2.14
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
requests==2.10.0
rply==0.7.4
sas==0.1.2
setuptools-bower==0.2.0
six==1.8.0
static3==0.5.1
status==0.2.5
stevedore==1.14.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
zope.component==4.2.2
zope.deprecation==4.1.2
zope.event==4.2.0
zope.interface==4.1.3
zope.lifecycleevent==4.1.0
